fix(win7/virtio-snd): silence W4 format warning in VirtIoSndSetupQueues
Cast numQueues/queue-count to UINT to match %u in VIRTIOSND_TRACE_ERROR.

+++ drivers/windows7/virtio-snd/src/virtiosnd_hw.c
@@ -84,8 +84,8 @@ static NTSTATUS VirtIoSndSetupQueues(_Inout_ PVIRTIOSND_DEVICE_EXTENSION Dx)
         if (numQueues < (USHORT)VIRTIOSND_QUEUE_COUNT) {
             VIRTIOSND_TRACE_ERROR(
                 "device exposes %u queues (< %u required by contract v1)\n",
-                (ULONG)numQueues,
-                (ULONG)VIRTIOSND_QUEUE_COUNT);
+                (UINT)numQueues,
+                (UINT)VIRTIOSND_QUEUE_COUNT);
             return STATUS_DEVICE_CONFIGURATION_ERROR;
         }
     }
